+++ test/client_handles_bad_connections_test.dart
@@ -16,7 +16,8 @@ class TestClient extends grpc.Client {
       (int value) => [value],
       (List<int> value) => value[0]);
 
-  TestClient(ClientChannel channel) : super(channel);
+  TestClient(super.channel);
+
   grpc.ResponseStream<int> stream(int request, {grpc.CallOptions? options}) {
     return $createStreamingCall(_$stream, Stream.value(request),
         options: options);
@@ -42,9 +43,13 @@ class TestService extends grpc.Service {
 class FixedConnectionClientChannel extends ClientChannelBase {
   final Http2ClientConnection clientConnection;
   List<grpc.ConnectionState> states = <grpc.ConnectionState>[];
+
   FixedConnectionClientChannel(this.clientConnection) {
-    clientConnection.onStateChanged = (c) => states.add(c.state);
+    onConnectionStateChanged.listen((state) {
+      states.add(state);
+    });
   }
+
   @override
   ClientConnection createConnection() => clientConnection;
 }
@@ -53,7 +58,7 @@ Future<void> main() async {
   testTcpAndUds('client reconnects after the connection gets old',
       (address) async {
     // client reconnect after a short delay.
-    final server = grpc.Server([TestService()]);
+    final server = grpc.Server.create(services: [TestService()]);
     await server.serve(address: address, port: 0);
 
     final channel = FixedConnectionClientChannel(Http2ClientConnection(
@@ -78,7 +83,7 @@ Future<void> main() async {
 
   testTcpAndUds('client reconnects when stream limit is used', (address) async {
     // client reconnect after setting stream limit.
-    final server = grpc.Server([TestService()]);
+    final server = grpc.Server.create(services: [TestService()]);
     await server.serve(
         address: address,
         port: 0,
@@ -89,8 +94,9 @@ Future<void> main() async {
         server.port!,
         grpc.ChannelOptions(credentials: grpc.ChannelCredentials.insecure())));
     final states = <grpc.ConnectionState>[];
-    channel.clientConnection.onStateChanged =
-        (Http2ClientConnection connection) => states.add(connection.state);
+    channel.onConnectionStateChanged.listen((state) {
+      states.add(state);
+    });
     final testClient = TestClient(channel);
 
     await Future.wait(<Future>[