+++ test/round_trip_test.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:grpc/grpc.dart';
-import 'package:grpc/service_api.dart' as api;
 import 'package:grpc/src/client/channel.dart' hide ClientChannel;
 import 'package:grpc/src/client/connection.dart';
 import 'package:grpc/src/client/http2_connection.dart';
@@ -15,7 +14,8 @@ class TestClient extends Client {
   static final _$stream = ClientMethod<int, int>('/test.TestService/stream',
       (int value) => [value], (List<int> value) => value[0]);
 
-  TestClient(api.ClientChannel channel) : super(channel);
+  TestClient(super.channel);
+
   ResponseStream<int> stream(int request, {CallOptions? options}) {
     return $createStreamingCall(_$stream, Stream.value(request),
         options: options);
@@ -23,10 +23,12 @@ class TestClient extends Client {
 }
 
 class TestService extends Service {
+  final String? expectedAuthority;
+
   @override
   String get $name => 'test.TestService';
 
-  TestService() {
+  TestService({this.expectedAuthority}) {
     $addMethod(ServiceMethod<int, int>('stream', stream, false, true,
         (List<int> value) => value[0], (int value) => [value]));
   }
@@ -35,10 +37,18 @@ class TestService extends Service {
   static const requestInfiniteStream = 2;
 
   Stream<int> stream(ServiceCall call, Future request) async* {
+    checkMetadata(call.clientMetadata);
+
     final isInfinite = 2 == await request;
     for (var i = 1; i <= 3 || isInfinite; i++) {
       yield i;
       await Future.delayed(Duration(milliseconds: 100));
+    }
+  }
+
+  void checkMetadata(Map<String, String>? metadata) {
+    if (expectedAuthority != null) {
+      expect(metadata![':authority'], equals(expectedAuthority));
     }
   }
 }
@@ -50,12 +60,30 @@ class TestServiceWithOnMetadataException extends TestService {
   }
 }
 
+class TestServiceWithGrpcError extends TestService {
+  @override
+  Stream<int> stream(ServiceCall call, Future request) async* {
+    throw GrpcError.custom(
+      StatusCode.internal,
+      'This error should contain trailers',
+      null,
+      null,
+      {
+        'key1': 'value1',
+        'key2': 'value2',
+      },
+    );
+  }
+}
+
 class FixedConnectionClientChannel extends ClientChannelBase {
   final Http2ClientConnection clientConnection;
   List<ConnectionState> states = <ConnectionState>[];
+
   FixedConnectionClientChannel(this.clientConnection) {
-    clientConnection.onStateChanged = (c) => states.add(c.state);
+    onConnectionStateChanged.listen((state) => states.add(state));
   }
+
   @override
   ClientConnection createConnection() => clientConnection;
 }
@@ -63,7 +91,24 @@ class FixedConnectionClientChannel extends ClientChannelBase {
 Future<void> main() async {
   testTcpAndUds('round trip insecure connection', (address) async {
     // round trip test of insecure connection.
-    final server = Server([TestService()]);
+    final server = Server.create(services: [TestService()]);
+    await server.serve(address: address, port: 0);
+
+    final channel = FixedConnectionClientChannel(Http2ClientConnection(
+      address,
+      server.port!,
+      ChannelOptions(credentials: ChannelCredentials.insecure()),
+    ));
+    final testClient = TestClient(channel);
+    expect(await testClient.stream(TestService.requestFiniteStream).toList(),
+        [1, 2, 3]);
+    server.shutdown();
+  });
+
+  testUds('UDS provides valid default authority', (address) async {
+    // round trip test of insecure connection.
+    final server =
+        Server.create(services: [TestService(expectedAuthority: 'localhost')]);
     await server.serve(address: address, port: 0);
 
     final channel = FixedConnectionClientChannel(Http2ClientConnection(
@@ -79,8 +124,10 @@ Future<void> main() async {
 
   testTcpAndUds('round trip with outgoing and incoming compression',
       (address) async {
-    final server = Server(
-        [TestService()], const [], CodecRegistry(codecs: const [GzipCodec()]));
+    final server = Server.create(
+      services: [TestService()],
+      codecRegistry: CodecRegistry(codecs: const [GzipCodec()]),
+    );
     await server.serve(address: address, port: 0);
 
     final channel = FixedConnectionClientChannel(Http2ClientConnection(
@@ -103,7 +150,7 @@ Future<void> main() async {
 
   testTcpAndUds('round trip secure connection', (address) async {
     // round trip test of secure connection.
-    final server = Server([TestService()]);
+    final server = Server.create(services: [TestService()]);
     await server.serve(
         address: address,
         port: 0,
@@ -126,7 +173,8 @@ Future<void> main() async {
   });
 
   test('exception in onMetadataException', () async {
-    final server = Server([TestServiceWithOnMetadataException()]);
+    final server =
+        Server.create(services: [TestServiceWithOnMetadataException()]);
     await server.serve(address: 'localhost', port: 0);
 
     final channel = FixedConnectionClientChannel(Http2ClientConnection(
@@ -142,7 +190,7 @@ Future<void> main() async {
   });
 
   test('cancellation of streaming subscription propagates properly', () async {
-    final server = Server([TestService()]);
+    final server = Server.create(services: [TestService()]);
     await server.serve(address: 'localhost', port: 0);
 
     final channel = FixedConnectionClientChannel(Http2ClientConnection(
@@ -153,6 +201,32 @@ Future<void> main() async {
     final testClient = TestClient(channel);
     expect(await testClient.stream(TestService.requestInfiniteStream).first, 1);
     await channel.shutdown();
+    await server.shutdown();
+  });
+
+  test('trailers on server GrpcError', () async {
+    final server = Server.create(services: [TestServiceWithGrpcError()]);
+    await server.serve(address: 'localhost', port: 0);
+
+    final channel = FixedConnectionClientChannel(Http2ClientConnection(
+      'localhost',
+      server.port!,
+      ChannelOptions(credentials: ChannelCredentials.insecure()),
+    ));
+    final testClient = TestClient(channel);
+    await expectLater(
+      testClient.stream(TestService.requestFiniteStream).toList(),
+      throwsA(predicate<GrpcError>((e) {
+        final trailers = e.trailers;
+        if (trailers == null || trailers.length != 2) return false;
+        final entries = trailers.entries.toList();
+        final isOk = entries[0].key == 'key1' &&
+            entries[0].value == 'value1' &&
+            entries[1].key == 'key2' &&
+            entries[1].value == 'value2';
+        return isOk;
+      })),
+    );
     await server.shutdown();
   });
 }